make preparedTxnState a separate class
